+++ android/src/main/kotlin/com/n1k3c/link_preview/LinkPreviewPlugin.kt
@@ -1,21 +1,20 @@
 package com.n1k3c.link_preview
 
+import com.leocardz.link.preview.library.LinkPreviewCallback
+import com.leocardz.link.preview.library.SourceContent
+import com.leocardz.link.preview.library.TextCrawler
+import io.flutter.plugin.common.EventChannel
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
 import io.flutter.plugin.common.PluginRegistry.Registrar
-import com.leocardz.link.preview.library.SourceContent
-import com.leocardz.link.preview.library.LinkPreviewCallback
-import com.leocardz.link.preview.library.TextCrawler
-import io.flutter.plugin.common.EventChannel
-import java.util.*
 
 
 private const val STATE = "state"
-private const val STATE_LOADING = "loading"
 private const val STATE_SUCCESS = "success"
-private const val STATE_ERROR = "error"
+private const val STATE_PARSING_ERROR = "parsing_error"
+private const val STATE_WRONG_URL = "wrong_url_error"
 
 private const val FIELD_TITLE = "title"
 private const val FIELD_DESC = "description"
@@ -28,26 +27,33 @@ private const val FIELD_IMAGE = "image"
 
 private const val ERROR_TYPE = "Link preview error"
 
-class LinkPreviewPlugin : EventChannel.StreamHandler {
+class LinkPreviewPlugin : MethodCallHandler {
+
     companion object {
         @JvmStatic
         fun registerWith(registrar: Registrar) {
-            val eventChannel = EventChannel(registrar.messenger(), "link_preview_events")
-            eventChannel.setStreamHandler(LinkPreviewPlugin())
+            val channel = MethodChannel(registrar.messenger(), "link_preview_channel")
+            channel.setMethodCallHandler(LinkPreviewPlugin())
         }
     }
 
-    override fun onListen(arguments: Any, event: EventChannel.EventSink) {
+    override fun onMethodCall(call: MethodCall, result: Result) {
+        if (call.method.equals("previewLink")) {
+            val url: String? = call.argument("url")
 
+            previewLink(url, result)
+        } else {
+            result.notImplemented()
+        }
+    }
+
+    private fun previewLink(url: String?, result: Result) {
         val data = mutableMapOf<String, String>()
 
         val textCrawler = TextCrawler()
 
         val linkPreviewCallback = object : LinkPreviewCallback {
-            override fun onPre() {
-                data[STATE] = STATE_LOADING
-                event.success(data)
-            }
+            override fun onPre() {}
 
             override fun onPos(sourceContent: SourceContent, b: Boolean) {
                 if (sourceContent.isSuccess) {
@@ -61,23 +67,19 @@ class LinkPreviewPlugin : EventChannel.StreamHandler {
                     data[FIELD_ROW] = sourceContent.raw
                     data[FIELD_HTML_CODE] = sourceContent.htmlCode
 
-                    event.success(data)
-                    event.endOfStream()
+                    result.success(data)
                 } else {
-                    event.error(ERROR_TYPE, "Parsing URL error. Check your URL for typos and/or your connection", "")
-                    event.endOfStream()
+                    data[STATE] = STATE_PARSING_ERROR
+                    result.success(data)
                 }
             }
         }
 
-        val url: String? = arguments.toString()
-
-        if (url.isNullOrBlank()) event.error(ERROR_TYPE, "URL is blank", "")
+        if (url.isNullOrBlank()) {
+            data[STATE] = STATE_WRONG_URL
+            result.success(data)
+        }
 
         textCrawler.makePreview(linkPreviewCallback, url)
-    }
-
-    override fun onCancel(p0: Any?) {
-
     }
 }