+++ lib/link_preview.dart
@@ -1,36 +1,28 @@
 import 'dart:async';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
 class LinkPreview {
-  static const EventChannel eventChannel = EventChannel('link_preview_events');
+  static const MethodChannel _methodChannel =
+      MethodChannel('link_preview_channel');
 
-  static Future<String> getPreview(String url,
-      {@required Function(PreviewResponse model) onData,
-      @required Function onError}) async {
-    return eventChannel
-        .receiveBroadcastStream(url)
-        .listen(
-          (dynamic event) => _mapData(event, onData),
-          onError: (dynamic error) => onError(error),
-        )
-        .asFuture();
+  static Future<PreviewResponse> getPreview(String url) async {
+    Map<String, dynamic> args = <String, dynamic>{};
+    args.putIfAbsent("url", () => url);
+
+    var result = await _methodChannel.invokeMethod("previewLink", args);
+    PreviewResponse previewResponse = _mapResultToResponse(result);
+    return previewResponse;
   }
 
-  static _mapData(dynamic event, Function(PreviewResponse model) onData) {
-    var data = Map<String, dynamic>.from(event);
+  static PreviewResponse _mapResultToResponse(dynamic result) {
+    var data = Map<String, dynamic>.from(result);
 
-    PreviewStatus status;
+    PreviewResponse previewResponse;
 
-    if (data['state'] == 'loading') {
-      status = PreviewStatus.loading;
-      PreviewResponse previewModel = PreviewResponse(status);
-      onData(previewModel);
-    } else {
-      status = PreviewStatus.complete;
-      PreviewResponse previewModel = PreviewResponse(
-        status,
+    if (data['state'] == 'success') {
+      previewResponse = PreviewResponse(
+        PreviewStatus.success,
         title: data['title'],
         description: data['description'],
         image: data['image'],
@@ -39,10 +31,16 @@ class LinkPreview {
         cannonicalUrl: data['cannonical_url'],
         row: data['row'],
         htmlCode: data['html_code'],
-
       );
-      onData(previewModel);
+    } else if (data['state'] == 'wrong_url_error') {
+      previewResponse = PreviewResponse(PreviewStatus.wrongUrlError);
+    } else if (data['state'] == 'parsing_error') {
+      previewResponse = PreviewResponse(PreviewStatus.parsingError);
+    } else {
+      previewResponse = PreviewResponse(PreviewStatus.otherError);
     }
+
+    return previewResponse;
   }
 }
 
@@ -68,4 +66,4 @@ class PreviewResponse {
       this.htmlCode});
 }
 
-enum PreviewStatus { loading, complete }
+enum PreviewStatus { success, wrongUrlError, parsingError, otherError }